Implemented the mvvm pattern for the ownStudies fragment
Because:
- All parts of the app should follow the mvvm pattern

Added a viewModel,repository and listener and renamed some classes
