+++ app/src/main/java/com/example/vpmanager/interfaces/StudyMetaDataListener.java
@@ -1,0 +1,9 @@
+package com.example.vpmanager.interfaces;
+
+import com.example.vpmanager.models.StudyMetaInfoModel;
+
+import java.util.ArrayList;
+
+public interface StudyMetaDataListener {
+    void onStudyMetaDataReady(ArrayList<StudyMetaInfoModel> studyMetaInfosArrayList);
+}
